Add predictGLM() function
In several places the GLM notes went through the grunt work of
calculating CIs for GLMs by back transforming the CIs on the linear
predictor scale since the built in predict.glm function doesn't provide
CIs. The s20x package does contain a predictCount function for this, but
it only works for Poisson models. So, I upgraded it to predictGLM to get
CIs for both Poisson and binomial models on either linear predictor or
response scale. I added use of predictGLM to the notes so that they
would just see the grunt work on one occasion and then use predictGLM
after that.