Fix summary1way() for R 4+ due to stringsAsFactors change
dummy.coef() no longer has the desired effect when a model has a
term that is of type = "character".  Small workaround for that
to use predict() to get the same numbers.

Includes a minor touch up to avoid rerunning split() where it had
already been run.

onewayPlot() also required a fix due to xlim's, so tweak to
calculate the xlims properly in a string-compliant manner.

Additionally, multipleComp would stop() due to
estimateConstrasts1() not allowing character vectors, it works
perfectly adding is.character() to the condition, so fixed.

Tested against CS11.1 and appears to work fine.